# [GraphQL] Path Checking Performance Fix