--- a/benchmark/sirun/plugin-graphql/index.js
+++ b/benchmark/sirun/plugin-graphql/index.js
@@ -4,7 +4,7 @@ const semver = require('semver')
 
 // TODO: benchmark the tracer as well but for now it's just too slow
 // if (Number(process.env.WITH_TRACER)) {
-//   require('../../..').init().use('graphql', { depth: 0 })
+require('../../..').init().use('graphql', { depth: 3, collapse: true })
 // }
 
 if (Number(process.env.WITH_ASYNC_HOOKS)) {
@@ -40,5 +40,6 @@ const source = `
 const variableValues = { who: 'world' }
 
 for (let i = 0; i < 5; i++) {
-  graphql.graphql({ schema, source, variableValues })
+  console.time(`Execution Time ${i}`)
+  graphql.graphql({ schema, source, variableValues }).then(() => console.timeEnd(`Execution Time ${i}`))
 }