--- a/benchmark/sirun/plugin-graphql/index.js
+++ b/benchmark/sirun/plugin-graphql/index.js
@@ -3,9 +3,9 @@
 const semver = require('semver')
 
 // TODO: benchmark the tracer as well but for now it's just too slow
-// if (Number(process.env.WITH_TRACER)) {
-require('../../..').init().use('graphql', { depth: 3, collapse: true })
-// }
+if (Number(process.env.WITH_TRACER)) {
+  require('../../..').init().use('graphql', { depth: 3, collapse: false })
+}
 
 if (Number(process.env.WITH_ASYNC_HOOKS)) {
   const hook = semver.satisfies(process.versions.node, '>=14.5 || ^12.19.0')
@@ -40,6 +40,5 @@ const source = `
 const variableValues = { who: 'world' }
 
 for (let i = 0; i < 5; i++) {
-  console.time(`Execution Time ${i}`)
-  graphql.graphql({ schema, source, variableValues }).then(() => console.timeEnd(`Execution Time ${i}`))
+  graphql.graphql({ schema, source, variableValues })
 }