--- a/benchmark/sirun/plugin-graphql/index.js
+++ b/benchmark/sirun/plugin-graphql/index.js
@@ -4,7 +4,15 @@ const semver = require('semver')
 
 // TODO: benchmark the tracer as well but for now it's just too slow
 if (Number(process.env.WITH_TRACER)) {
-  require('../../..').init().use('graphql', { depth: 3, collapse: false })
+  const tracer = require('../../..').init()
+
+  // Note: depth must be an integer >= 0, and collapse either 0 or 1 (true or false)
+  if (Number(process.env.WITH_DEPTH)) {
+    tracer.use('graphql', { depth: +process.env.WITH_DEPTH })
+  } else if (process.env.WITH_DEPTH_AND_COLLAPSE) {
+    const [depth, collapse] = process.env.WITH_DEPTH_AND_COLLAPSE.split(',')
+    tracer.use('graphql', { depth: +depth, collapse: !!+collapse })
+  }
 }
 
 if (Number(process.env.WITH_ASYNC_HOOKS)) {