--- a/benchmark/sirun/plugin-graphql/meta.json
+++ b/benchmark/sirun/plugin-graphql/meta.json
@@ -8,6 +8,21 @@
     "with-async-hooks": {
       "baseline": "control",
       "env": { "WITH_ASYNC_HOOKS": "1" }
+    },
+    "with-depth-off": {
+      "baseline": "control",
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH": "0"}
+    },"with-depth-on-max": {
+      "baseline": "control",
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH": "4"}
+    },
+    "with-depth-and-collapse-off": {
+      "baseline": "control",
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH_AND_COLLAPSE": "4,0"}
+    },
+    "with-depth-and-collapse-on": {
+      "baseline": "control",
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH_AND_COLLAPSE": "4,1"}
     }
   }
 }