--- a/benchmark/sirun/plugin-graphql/meta.json
+++ b/benchmark/sirun/plugin-graphql/meta.json
@@ -11,18 +11,22 @@
     },
     "with-depth-off": {
       "baseline": "control",
-      "env": { "WITH_TRACER": "1", "WITH_DEPTH": "0"}
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH": "0"},
+      "cachegrind": false
     },"with-depth-on-max": {
       "baseline": "control",
-      "env": { "WITH_TRACER": "1", "WITH_DEPTH": "4"}
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH": "4"},
+      "cachegrind": false
     },
     "with-depth-and-collapse-off": {
       "baseline": "control",
-      "env": { "WITH_TRACER": "1", "WITH_DEPTH_AND_COLLAPSE": "4,0"}
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH_AND_COLLAPSE": "4,0"},
+      "cachegrind": false
     },
     "with-depth-and-collapse-on": {
       "baseline": "control",
-      "env": { "WITH_TRACER": "1", "WITH_DEPTH_AND_COLLAPSE": "4,1"}
+      "env": { "WITH_TRACER": "1", "WITH_DEPTH_AND_COLLAPSE": "4,1"},
+      "cachegrind": false
     }
   }
 }