### What does this PR do?
Improves path-checking performance

### Motivation
Fixes #2159 
The `hasLikePath` function was slow, likely due to the use of iterative application of a regex test. This PR attempts to resolve it through means other than regex, starting with a similar iterative approach, but doing more basic logic checks.

### Performance Measurement

All these metrics come from the benchmark structure and schema in `./benchmark/sirun/plugin-graphql`, which runs a stress test of 5 rounds against a runtime-heavy schema.

5 runs w/o tracer:

388.651ms, 386.741ms, 386.139ms, 385.662ms, 382.978ms

5 runs w/ tracer, before final changes (using intermediate commit, which has iterative approach)

**collapse set to true**: 1.702s, 1.682s, 1.680s, 1.678s, 1.677s
**collapse set to false**: 17:19.589, 17:19.555, 17:19.553, 17:19.551, 17:19.550 (all in m:ss.mmm)

5 runs w/ tracer, after final changes (cached approach)

**collapse set to true**: 1.889s, 1.869s, 1.866s, 1.864s, 1.864s
**collapse set to false**: 2.931s, 2.910s, 2.907s, 2.905s, 2.904s

Note that these results *are on* a stress test, so an individual run would be faster as well, but the stress test shows increased performance on a consistent basis.